mark down success

+++ src/main/java/com/zym/community/mapper/UserMapper.java
@@ -20,10 +20,11 @@ public interface UserMapper {
     User findByToken(@Param("token") String token);
 
     @Select("select * from user where id = #{id}")
-    User findById(@RequestParam("id") Integer creator);
+    User findById(@RequestParam("id") long creator);
 
     @Select("select * from user where account_id = #{accountId}")
     User findByAccountId(String accountId);
+
 
     @Update("update user set name=#{name},token=#{token},gmt_modified=#{gmtModified},avatar_url=#{avatarUrl} where id = #{id}")
     void update(User user);
